Made fill_map.c and moved some functions to it. Streamlined error handling
