+++ src/main_bonus.c
@@ -6,7 +6,7 @@
 /*   By: pikkak <pikkak@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/12 15:57:40 by tsaari            #+#    #+#             */
-/*   Updated: 2024/10/25 20:56:37 by pikkak           ###   ########.fr       */
+/*   Updated: 2024/10/25 21:40:52 by pikkak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -68,32 +68,28 @@ void free_before_map(t_data *data)
 	free(data->scene.so);
 	free(data->scene.no);
 	free(data->scene.we);
-	//free(data);
-	ft_error(data, "File cannot be opened");
+	ft_error(data, ERR_OPEN);
 }
 
 void parse(t_data *data)
 {
 	t_check *check;
-
-	check = (t_check *)malloc(sizeof (t_check));//do we need to allocate this to the heap?
+	
+	check = (t_check *)malloc(sizeof (t_check));
 	init_check(check);
 	if (check_filetype(data->file, ".cub") != 0)
 		ft_error(data, ERR_INFILE);
 	check_and_set_texttr_and_col_lines(data, check);
 	check_map_lines(data, check);
-	data->fd = open(data->file, O_RDONLY);
+	data->fd = open(data->file, O_RDONLY);//Open the file again from the beginning
 	if (data->fd < 0)
 		free_before_map(data);
 	read_file_for_longest_and_lines(data, check);
-	data->scene.rows = check->map_lines + 10;
-	data->scene.cols = check->longest_line + 10;
 	allocate_map(data);
 	init_map(data);
-	close(data->fd);
 	data->fd = open(data->file, O_RDONLY);
 	if (data->fd < 0)
-		ft_free_data_and_error(data, "File cannot be opened");
+		ft_free_data_and_error(data, ERR_OPEN);
 	fill_map(data, check);
 	flood_fill(data, (t_point){data->scene.cols, data->scene.rows}, (t_point){0,0}, 32);
 	fill_maze_if_spaces(data);
@@ -109,19 +105,19 @@ int	main(int argc, char **argv)
 
 	if (argc != 2)
 	{
-		write(2, "Invalid amount of arguments", 27);
+		write(2, ERR_ARG, 27);
 		return (0);
 	}
 	data = (t_data *)malloc(sizeof(t_data));
 	if (!data)
-		ft_error(data, "malloc error");
+		ft_error(data, ERR_MALLOC);
 	else
 	{
 		init_data(data, argv);
 		parse(data);
 		data->m = mlx_init(data->s_width, data->s_height, "Cub3D", false);
 		if (!data->m)
-			ft_free_data_and_error(data, "MLX error");
+			ft_free_data_and_error(data, ERR_MLX);
 		get_textures(data);
 		mlx_loop_hook(data->m, my_keyhook, data);
 		mlx_key_hook(data->m, &my_keyhook2, data);
