+++ src/parsing/parse_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse_utils.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kkauhane <kkauhane@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: pikkak <pikkak@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/01 16:28:19 by kkauhane          #+#    #+#             */
-/*   Updated: 2024/10/24 14:53:14 by kkauhane         ###   ########.fr       */
+/*   Updated: 2024/10/25 21:48:04 by pikkak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,38 +64,18 @@ void allocate_map(t_data *data)
 	data->scene.map = malloc(data->scene.rows * sizeof(int*));
 	if (!data->scene.map)
 	{
-		ft_free_data_and_error(data, "malloc error");
+		ft_free_data_and_error(data, ERR_MALLOC);
 	}
 	while (row < data->scene.rows)
 	{
 		data->scene.map[row] = malloc(data->scene.cols * sizeof(int));
 		if (!data->scene.map[row])
 		{
-			ft_free_data_and_error(data, "malloc error");
+			ft_free_data_and_error(data, ERR_MALLOC);
 		}
 		row++;
 	}
 }
 
-
-void	flood_fill(t_data *data, t_point size, t_point cur, int to_fill)
-{
-	int **matrix = data->scene.map;
-
-	if (cur.y < 0 || cur.y >= size.y || cur.x < 0 || cur.x >= size.x)
-		return;
-	if (matrix[cur.y][cur.x] != 1 && matrix[cur.y][cur.x] != 32 && matrix[cur.y][cur.x] != -1)
-	{
-		ft_free_data_and_error(data, "invalid file");
-		return;
-	}
-	if (matrix[cur.y][cur.x] == 1 || matrix[cur.y][cur.x] == -1)
-		return;
-	matrix[cur.y][cur.x] = -1;
-	flood_fill(data, size, (t_point){cur.x - 1, cur.y}, to_fill);
-	flood_fill(data, size, (t_point){cur.x + 1, cur.y}, to_fill);
-	flood_fill(data, size, (t_point){cur.x, cur.y - 1}, to_fill);
-	flood_fill(data, size, (t_point){cur.x, cur.y + 1}, to_fill);
-}
